fixing test imports
